+++ lib/runner-helpers/handle-multiple-processes.js
@@ -2,40 +2,32 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 var process = require('suman-browser-polyfills/modules/process');
 var global = require('suman-browser-polyfills/modules/global');
-var cp = require("child_process");
-var fs = require("fs");
 var path = require("path");
-var util = require("util");
 var EE = require("events");
-var semver = require("semver");
-var merge = require('lodash.merge');
 var shuffle = require('lodash.shuffle');
 var suman_events_1 = require("suman-events");
 var su = require("suman-utils");
-var async = require("async");
 var noFilesFoundError = require('../helpers/no-files-found-error');
 var chalk = require("chalk");
 var _suman = global.__suman = (global.__suman || {});
 var resultBroadcaster = _suman.resultBroadcaster = (_suman.resultBroadcaster || new EE());
-var runnerUtils = require('./runner-utils');
 var socket_cp_hash_1 = require("./socket-cp-hash");
-var getTapParser = require('./handle-tap').getTapParser;
-var getTapJSONParser = require('./handle-tap-json').getTapJSONParser;
 var constants = require('../../config/suman-constants').constants;
-var debug = require('suman-debug')('s:runner');
+var transpile_queue_1 = require("./multi-process/transpile-queue");
+var add_to_transpile_queue_1 = require("./multi-process/add-to-transpile-queue");
 var multiple_process_each_on_exit_1 = require("./multiple-process-each-on-exit");
-var prepend_transform_1 = require("prepend-transform");
-var runChildPath = require.resolve(__dirname + '/run-child.js');
-var uuidV4 = require("uuid/v4");
+var run_file_1 = require("./multi-process/run-file");
 exports.makeHandleMultipleProcesses = function (runnerObj, tableRows, messages, forkedCPs, handleMessage, beforeExitRunOncePost, makeExit) {
     return function (runObj) {
+        _suman.startDateMillis = Date.now();
         process.stderr.setMaxListeners(runObj.files.length + 11);
         process.stdout.setMaxListeners(runObj.files.length + 11);
         var logsDir = _suman.sumanConfig.logsDir || _suman.sumanHelperDirRoot + '/logs';
         var sumanCPLogs = path.resolve(logsDir + '/runs/');
         var f = path.resolve(sumanCPLogs + '/' + _suman.timestamp + '-' + _suman.runId);
-        _suman.startDateMillis = Date.now();
-        var sumanOpts = _suman.sumanOpts, sumanConfig = _suman.sumanConfig, maxProcs = _suman.maxProcs, projectRoot = _suman.projectRoot;
+        var onExitFn = multiple_process_each_on_exit_1.makeOnExitFn(runnerObj, tableRows, messages, forkedCPs, beforeExitRunOncePost, makeExit, transpileQueue);
+        var runFile = run_file_1.makeRunFile(onExitFn);
+        var sumanOpts = _suman.sumanOpts, sumanConfig = _suman.sumanConfig, projectRoot = _suman.projectRoot;
         var waitForAllTranformsToFinish = sumanOpts.wait_for_all_transforms;
         if (waitForAllTranformsToFinish) {
             _suman.log('waitForAllTranformsToFinish => ', chalk.magenta(waitForAllTranformsToFinish));
@@ -43,27 +35,7 @@ exports.makeHandleMultipleProcesses = function (runnerObj, tableRows, messages, 
         var args = ['--user-args', sumanOpts.user_args];
         var queuedTestFns = [];
         var failedTransformObjects = [];
-        var transpileQueue = async.queue(function (task, cb) {
-            task(function (err, file, shortFile, stdout, stderr, gd) {
-                if (err) {
-                    _suman.logError('tranpile error => ', err.stack || err);
-                    failedTransformObjects.push({ err: err, file: file, shortFile: shortFile, stdout: stdout, stderr: stderr });
-                    return;
-                }
-                setImmediate(function () {
-                    console.log(chalk.red('pushing file '), file);
-                    if (waitForAllTranformsToFinish) {
-                        queuedTestFns.push(function () {
-                            outer(file, shortFile, stdout, gd);
-                        });
-                    }
-                    else {
-                        outer(file, shortFile, stdout, gd);
-                    }
-                    cb(null);
-                });
-            });
-        }, 3);
+        var transpileQueue = transpile_queue_1.makeTranspileQueue(failedTransformObjects, runFile, queuedTestFns);
         if (waitForAllTranformsToFinish) {
             transpileQueue.drain = function () {
                 _suman.log('all transforms complete, beginning to run first set of tests.');
@@ -78,7 +50,6 @@ exports.makeHandleMultipleProcesses = function (runnerObj, tableRows, messages, 
                     'and will not be listening for websocket messages.'));
             }
         }
-        var handleBlocking = runnerObj.handleBlocking;
         if (_suman.usingLiveSumanServer) {
             args.push('--live_suman_server');
         }
@@ -89,357 +60,6 @@ exports.makeHandleMultipleProcesses = function (runnerObj, tableRows, messages, 
         }
         runnerObj.startTime = Date.now();
         var fileObjArray = su.removeSharedRootPath(files);
-        var sumanEnv = Object.assign({}, process.env, {
-            SUMAN_RUN_CHILD_STATIC_PATH: runChildPath,
-            SUMAN_CONFIG: JSON.stringify(sumanConfig),
-            SUMAN_OPTS: JSON.stringify(sumanOpts),
-            SUMAN_RUNNER: 'yes',
-            SUMAN_PROJECT_ROOT: projectRoot,
-            SUMAN_RUN_ID: _suman.runId,
-            SUMAN_RUNNER_TIMESTAMP: _suman.timestamp,
-            NPM_COLORS: process.env.NPM_COLORS || (sumanOpts.no_color ? 'no' : 'yes')
-        });
-        if (_suman.socketServerPort > 0) {
-            sumanEnv['SUMAN_SOCKETIO_SERVER_PORT'] = _suman.socketServerPort;
-        }
-        var execFile = path.resolve(__dirname + '/run-child.js');
-        var istanbulExecPath = _suman.istanbulExecPath || 'istanbul';
-        var isStdoutSilent = sumanOpts.stdout_silent || sumanOpts.silent;
-        var isStderrSilent = sumanOpts.stderr_silent || sumanOpts.silent;
-        var inheritTransformStdio = sumanOpts.inherit_all_stdio || sumanOpts.inherit_transform_stdio || process.env.SUMAN_INHERIT_STDIO;
-        fileObjArray.forEach(function (fileShortAndFull) {
-            var uuid = String(uuidV4());
-            var file = fileShortAndFull[0];
-            var shortFile = fileShortAndFull[1];
-            var filePathFromProjectRoot = fileShortAndFull[2];
-            var basename = file.length > 28 ? ' ' + String(file).substring(Math.max(0, file.length - 28)) + ' ' : file;
-            var m = String(basename).match(/\//g);
-            if (m && m.length > 1) {
-                var arr = String(basename).split('');
-                var i = 0;
-                while (arr[i] !== '/') {
-                    arr.shift();
-                }
-                basename = arr.join('');
-            }
-            tableRows[shortFile] = {
-                actualExitCode: null,
-                shortFilePath: shortFile,
-                tableData: null,
-                defaultTableData: {
-                    SUITES_DESIGNATOR: basename
-                }
-            };
-            var gd = socket_cp_hash_1.ganttHash[uuid] = {
-                uuid: uuid,
-                fullFilePath: String(file),
-                shortFilePath: String(shortFile),
-                filePathFromProjectRoot: String(filePathFromProjectRoot),
-            };
-            var tr = (sumanOpts.no_transpile !== true) && runnerUtils.findPathOfTransformDotSh(file);
-            if (tr) {
-                _suman.log(chalk.bgWhite.underline('Suman has found a @transform.sh file => '), chalk.bold(tr));
-                transpileQueue.push(function (cb) {
-                    su.makePathExecutable(tr, function (err) {
-                        if (err) {
-                            return cb(err);
-                        }
-                        gd.transformStartDate = Date.now();
-                        var k = cp.spawn(tr, [], {
-                            cwd: projectRoot,
-                            env: Object.assign({}, process.env, {
-                                SUMAN_TEST_PATHS: JSON.stringify([file]),
-                                SUMAN_CHILD_TEST_PATH: file
-                            })
-                        });
-                        console.log(chalk.red('startttting..'));
-                        k.once('error', cb);
-                        k.stderr.setEncoding('utf8');
-                        k.stdout.setEncoding('utf8');
-                        var ln = String(_suman.projectRoot).length;
-                        if (false) {
-                            var onError = function (e) {
-                                _suman.logError('\n', su.getCleanErrorString(e), '\n');
-                            };
-                            var temp = su.removePath(file, _suman.projectRoot);
-                            var onlyFile = String(temp).replace(/\//g, '.');
-                            var logfile = path.resolve(f + '/' + onlyFile + '.log');
-                            var fileStrm = fs.createWriteStream(logfile);
-                            k.stderr.pipe(fileStrm).once('error', onError);
-                            k.stdout.pipe(fileStrm).once('error', onError);
-                        }
-                        if (inheritTransformStdio) {
-                            var onError = function (e) {
-                                _suman.logError('\n', su.getCleanErrorString(e), '\n');
-                            };
-                            var stderrPrepend = " [" + chalk.red('transform process stderr:') + " " + chalk.red.bold(String(file.slice(ln))) + "] ";
-                            k.stderr.pipe(prepend_transform_1.default(stderrPrepend, { omitWhitespace: true })).once('error', onError).pipe(process.stderr);
-                            var stdoutPrepend = " [" + chalk.yellow('transform process stdout:') + " " + chalk.gray.bold(String(file.slice(ln))) + "] ";
-                            k.stdout.pipe(prepend_transform_1.default(stdoutPrepend)).once('error', onError).pipe(process.stdout);
-                        }
-                        var stdout = '';
-                        k.stdout.on('data', function (data) {
-                            stdout += data;
-                        });
-                        var stderr = '';
-                        k.stderr.on('data', function (data) {
-                            stderr += data;
-                        });
-                        k.once('exit', function (code) {
-                            console.log(chalk.red('exxxxited.'));
-                            gd.transformEndDate = Date.now();
-                            if (code > 0) {
-                                cb(new Error("the @transform.sh process, for file " + file + ",\nexitted with non-zero exit code. :( \n                  \n To see the stderr, use --inherit-stdio."));
-                            }
-                            else {
-                                cb(null, file, shortFile, stdout, stderr, gd);
-                            }
-                        });
-                    });
-                });
-            }
-            else {
-                gd.transformStartDate = gd.transformEndDate = null;
-                gd.wasTransformed = false;
-                transpileQueue.unshift(function (cb) {
-                    setImmediate(function () {
-                        cb(null, file, shortFile, '', '', gd);
-                    });
-                });
-            }
-        });
-        var childId = 1;
-        var debugChildren = sumanOpts.debug_child || sumanOpts.inspect_child;
-        var inheritRunStdio = debugChildren || sumanOpts.inherit_stdio ||
-            sumanOpts.inherit_all_stdio || process.env.SUMAN_INHERIT_STDIO === 'yes';
-        var outer = function (file, shortFile, stdout, gd) {
-            var run = function () {
-                if (runnerObj.bailed) {
-                    if (sumanOpts.verbosity > 4) {
-                        _suman.log('"--bailed" option was passed and was tripped, ' +
-                            'no more child processes will be forked.');
-                    }
-                    return;
-                }
-                var argz = JSON.parse(JSON.stringify(args));
-                var execArgz = ['--expose-gc'];
-                if (sumanOpts.debug_child) {
-                    execArgz.push('--debug=' + (5303 + runnerObj.processId++));
-                    execArgz.push('--debug-brk');
-                }
-                if (sumanOpts.inspect_child) {
-                    if (semver.gt(process.version, '7.8.0')) {
-                        execArgz.push('--inspect-brk=' + (5303 + runnerObj.processId++));
-                    }
-                    else {
-                        execArgz.push('--inspect=' + (5303 + runnerObj.processId++));
-                        execArgz.push('--debug-brk');
-                    }
-                }
-                var execArgs;
-                if (execArgs = sumanOpts.exec_arg) {
-                    execArgs.forEach(function (n) {
-                        if (n) {
-                            execArgz.push(String(n).trim());
-                        }
-                    });
-                    String(execArgs).split(/S+/).forEach(function (n) {
-                        if (n) {
-                            execArgz.push('--' + String(n).trim());
-                        }
-                    });
-                }
-                var $execArgz = execArgz.filter(function (e, i) {
-                    if (execArgz.indexOf(e) !== i) {
-                        console.error('\n', chalk.yellow(' => Warning you have duplicate items in your exec args => '), '\n' + util.inspect(execArgz), '\n');
-                    }
-                    return true;
-                });
-                var n, hashbang = false;
-                var extname = path.extname(shortFile);
-                var $childId = childId++;
-                var childUuid = uuidV4();
-                var inherit = _suman.$forceInheritStdio ? 'inherit' : '';
-                if (inherit) {
-                    _suman.log('we are inheriting stdio of child, because of sumanception.');
-                }
-                var cpOptions = {
-                    detached: false,
-                    cwd: projectRoot,
-                    stdio: [
-                        'ignore',
-                        inherit || (isStdoutSilent ? 'ignore' : 'pipe'),
-                        inherit || (isStderrSilent ? 'ignore' : 'pipe'),
-                    ],
-                    env: Object.assign({}, sumanEnv, {
-                        SUMAN_CHILD_TEST_PATH: file,
-                        SUMAN_CHILD_TEST_PATH_TARGET: file,
-                        SUMAN_TRANSFORM_STDOUT: stdout,
-                        SUMAN_CHILD_ID: String($childId),
-                        SUMAN_CHILD_UUID: String($childId)
-                    })
-                };
-                var sh = runnerUtils.findPathOfRunDotSh(file);
-                if (sh) {
-                    _suman.log(chalk.bgWhite.underline('Suman has found a @run.sh file => '), chalk.bold(sh));
-                    cpOptions.cwd = projectRoot;
-                    try {
-                        fs.chmodSync(sh, 511);
-                    }
-                    catch (err) {
-                    }
-                    if (sumanOpts.coverage) {
-                        _suman.logWarning(chalk.magenta('coverage option was set to true, but we are running your tests via @run.sh.'));
-                        _suman.logWarning(chalk.magenta('so in this case, you will need to run your coverage call via @run.sh.'));
-                    }
-                    n = cp.spawn(sh, argz, cpOptions);
-                }
-                else {
-                    if ('.js' === extname) {
-                        if (sumanOpts.coverage) {
-                            var coverageDir = path.resolve(_suman.projectRoot + '/coverage/' + String(shortFile).replace(/\//g, '-'));
-                            n = cp.spawn(istanbulExecPath, ['cover', execFile, '--dir', coverageDir, '--'].concat(args), cpOptions);
-                        }
-                        else {
-                            argz.unshift(execFile);
-                            var argzz = $execArgz.concat(argz);
-                            n = cp.spawn('node', argzz, cpOptions);
-                        }
-                    }
-                    else {
-                        _suman.log("perl bash python or ruby file? '" + chalk.magenta(file) + "'");
-                        hashbang = true;
-                        n = cp.spawn(file, argz, cpOptions);
-                    }
-                }
-                socket_cp_hash_1.cpHash[$childId] = n;
-                if (!_suman.weAreDebugging) {
-                    n.to = setTimeout(function () {
-                        _suman.logError("Suman killed a child process because it timed out: '" + (n.fileName || n.filename) + "'.");
-                        n.kill('SIGINT');
-                        setTimeout(function () {
-                            n.kill('SIGKILL');
-                        }, 8000);
-                    }, constants.DEFAULT_CHILD_PROCESS_TIMEOUT);
-                }
-                n.testPath = file;
-                n.shortTestPath = shortFile;
-                forkedCPs.push(n);
-                n.on('message', function (msg) {
-                    _suman.logError('Suman runner does not handle standard Node.js IPC messages.');
-                });
-                n.on('error', function (err) {
-                    _suman.logError('error spawning child process => ', err.stack || err);
-                    if (hashbang) {
-                        console.error('\n');
-                        console.error(' => The supposed test script file with the following path may not have a hashbang => ');
-                        console.error(chalk.magenta.bold(file));
-                        console.error(' => A hashbang is necessary for non-.js files and when there is no accompanying @run.sh file.');
-                        console.error(' => Without a hashbang, Suman (and your OS) will not know how to run the file.');
-                        console.error(' => See sumanjs.org for more information.');
-                    }
-                });
-                if (n.stdio && n.stdout && n.stderr) {
-                    if (inherit) {
-                        _suman.logError('n.stdio is defined even though we are in sumanception territory.');
-                    }
-                    n.stdout.setEncoding('utf8');
-                    n.stderr.setEncoding('utf8');
-                    if (false && (sumanOpts.log_stdio_to_files || sumanOpts.log_stdout_to_files || sumanOpts.log_stderr_to_files)) {
-                        var onError = function (e) {
-                            _suman.logError('\n', su.getCleanErrorString(e), '\n');
-                        };
-                        var temp = su.removePath(file, _suman.projectRoot);
-                        var onlyFile = String(temp).replace(/\//g, '.');
-                        var logfile = path.resolve(f + '/' + onlyFile + '.log');
-                        var fileStrm = fs.createWriteStream(logfile);
-                        console.log('logFile => ', logfile);
-                        if (sumanOpts.log_stdio_to_files || sumanOpts.log_stderr_to_files) {
-                            n.stderr.pipe(fileStrm).once('error', onError);
-                        }
-                        if (sumanOpts.log_stdio_to_files || sumanOpts.log_stdout_to_files) {
-                            n.stdout.pipe(fileStrm).once('error', onError);
-                        }
-                    }
-                    if (inheritRunStdio) {
-                        var onError = function (e) {
-                            _suman.logError('\n', su.getCleanErrorString(e), '\n');
-                        };
-                        n.stdout.pipe(prepend_transform_1.default(chalk.cyan(' [suman child stdout] ')))
-                            .once('error', onError).pipe(process.stdout);
-                        n.stderr.pipe(prepend_transform_1.default(chalk.red.bold(' [suman child stderr] '), { omitWhitespace: true }))
-                            .once('error', onError).pipe(process.stderr);
-                    }
-                    if (true || sumanOpts.$useTAPOutput) {
-                        n.tapOutputIsComplete = false;
-                        n.stdout.pipe(getTapParser())
-                            .on('error', function (e) {
-                            _suman.logError('error parsing TAP output =>', su.getCleanErrorString(e));
-                        })
-                            .once('finish', function () {
-                            n.tapOutputIsComplete = true;
-                            process.nextTick(function () {
-                                n.emit('tap-output-is-complete', true);
-                            });
-                        });
-                        n.stdout.pipe(getTapJSONParser())
-                            .on('error', function (e) {
-                            _suman.logError('error parsing TAP JSON output =>', su.getCleanErrorString(e));
-                        });
-                    }
-                    n.stdio[2].setEncoding('utf-8');
-                    n.stdio[2].on('data', function (data) {
-                        var d = String(data).split('\n').filter(function (line) {
-                            return String(line).length;
-                        })
-                            .map(function (line) {
-                            return '[' + n.shortTestPath + '] ' + line;
-                        })
-                            .join('\n');
-                        _suman.sumanStderrStream.write('\n' + d);
-                        if (_suman.weAreDebugging) {
-                            console.log('pid => ', n.pid, 'stderr => ', d);
-                        }
-                    });
-                }
-                else {
-                    if (su.vgt(2)) {
-                        _suman.logWarning('Stdio object not available for child process.');
-                    }
-                }
-                n.dateStartedMillis = gd.startDate = Date.now();
-                n.once('exit', multiple_process_each_on_exit_1.default(n, runnerObj, tableRows, messages, forkedCPs, beforeExitRunOncePost, makeExit, gd, transpileQueue));
-            };
-            run.testPath = file;
-            run.shortTestPath = shortFile;
-            if (handleBlocking.runNext(run)) {
-                if (su.vgt(3)) {
-                    _suman.log(chalk.black('File has just started running =>'), chalk.grey.bold("'" + file + "'."));
-                }
-            }
-            else {
-                runnerObj.queuedCPs.push(run);
-                if (su.vgt(3)) {
-                    _suman.log(chalk.yellow('File must wait, and will run later =>'), chalk.grey.bold("'" + file + "'."));
-                }
-            }
-            if (waitForAllTranformsToFinish) {
-                if (forkedCPs.length < 1 && runnerObj.queuedCPs.length > 0) {
-                    throw new Error('Suman internal error => fatal start order algorithm error, ' +
-                        'please file an issue on Github, thanks.');
-                }
-                if (forkedCPs.length < 1) {
-                    noFilesFoundError(files);
-                }
-                else {
-                    var totalCount = forkedCPs.length + runnerObj.queuedCPs.length;
-                    var suites = totalCount === 1 ? 'suite' : 'suites';
-                    var processes = totalCount === 1 ? 'process' : 'processes';
-                    resultBroadcaster.emit(String(suman_events_1.events.RUNNER_INITIAL_SET), forkedCPs, processes, suites);
-                    var addendum = maxProcs < totalCount ? ' with no more than ' + maxProcs + ' running at a time.' : '';
-                    resultBroadcaster.emit(String(suman_events_1.events.RUNNER_OVERALL_SET), totalCount, processes, suites, addendum);
-                }
-            }
-        };
+        fileObjArray.forEach(add_to_transpile_queue_1.makeAddToTranspileQueue(f, transpileQueue, tableRows, socket_cp_hash_1.ganttHash, projectRoot));
     };
 };
