+++ lib/runner-helpers/make-handle-blocking.js
@@ -27,7 +27,8 @@ function default_1(order) {
                     return ended.every(function (item) {
                         return (String(item.testPath) !== String($item.testPath));
                     });
-                }).map(function (item) {
+                })
+                    .map(function (item) {
                     return '\n  ' + item.testPath;
                 });
                 if (startedButNotEnded.length > 0) {
@@ -39,11 +40,11 @@ function default_1(order) {
             }, initialTimeout += increaseEachTime);
         }, interval);
     }
-    function findQueuedCPsToStart(queuedCPsObj) {
+    var findQueuedCPsToStart = function (queuedCPsObj) {
         if (started.length - ended.length < maxProcs) {
             return queuedCPsObj.queuedCPs.pop();
         }
-    }
+    };
     return {
         runNext: function (fn) {
             if (started.length - ended.length < maxProcs) {
