+++ lib/runner-helpers/multi-process/add-to-transpile-queue.d.ts
@@ -1,0 +1,1 @@
+export declare const makeAddToTranspileQueue: (f: string, transpileQueue: any, tableRows: any, ganttHash: any, projectRoot: string) => (fileShortAndFull: string[][]) => void;
