+++ lib/runner-helpers/multi-process/run-file.d.ts
@@ -1,0 +1,2 @@
+import { IGanttData } from "../socket-cp-hash";
+export declare const makeRunFile: (runnerObj: any, args: any, runQueue: any, projectRoot: string, cpHash: any, forkedCPs: any, onExitFn: any) => (file: string, shortFile: string, stdout: string, gd: IGanttData) => void;
