+++ lib/runner-helpers/multi-process/run-queue.d.ts
@@ -1,0 +1,2 @@
+import async = require('async');
+export declare const makeRunQueue: () => async.AsyncQueue<{}>;
