+++ lib/runner-helpers/multi-process/transpile-queue.d.ts
@@ -1,0 +1,2 @@
+import async = require('async');
+export declare const makeTranspileQueue: (failedTransformObjects: any, outer: any, queuedTestFns: any) => async.AsyncQueue<Function>;
