+++ lib/runner-helpers/multiple-process-each-on-exit.d.ts
@@ -1,4 +1,4 @@
 import { IRunnerObj, ISumanChildProcess, ITableRows } from "suman-types/dts/runner";
 import { ISumanCPMessages } from "./handle-multiple-processes";
 import { IGanttData } from "./socket-cp-hash";
-export default function (n: ISumanChildProcess, runnerObj: IRunnerObj, tableRows: ITableRows, messages: Array<ISumanCPMessages>, forkedCPs: Array<ISumanChildProcess>, beforeExitRunOncePost: Function, makeExit: Function, gd: IGanttData, transpileQueue: any): (code: number, signal: number) => void;
+export declare const makeOnExitFn: (runnerObj: IRunnerObj, tableRows: ITableRows, messages: ISumanCPMessages[], forkedCPs: ISumanChildProcess[], beforeExitRunOncePost: Function, makeExit: Function, gd: IGanttData, transpileQueue: any) => (n: ISumanChildProcess) => (code: number, signal: number) => void;
