+++ lib/runner.js
@@ -25,8 +25,7 @@ var mapValues = require('lodash.mapvalues');
 var chalk = require("chalk");
 var a8b = require('ansi-256-colors'), fg = a8b.fg, bg = a8b.bg;
 var suman_events_1 = require("suman-events");
-var suman_utils_1 = require("suman-utils");
-var debug = require('suman-debug')('s:runner');
+var su = require("suman-utils");
 var _suman = global.__suman = (global.__suman || {});
 var integrant_injector_1 = require("./injection/integrant-injector");
 var suman_constants_1 = require("../config/suman-constants");
@@ -74,11 +73,11 @@ var beforeExitRunOncePost = make_before_exit_once_post_1.makeBeforeExit(runnerOb
     process.once('exit', onExit);
 }
 process.on('error', function (e) {
-    _suman.logError(chalk.magenta('Whoops! "error" event in runner process:') + " \n " + chalk.bold(suman_utils_1.default.getCleanErrorString(e)));
+    _suman.logError(chalk.magenta('Whoops! "error" event in runner process:') + " \n " + chalk.bold(su.getCleanErrorString(e)));
 });
 process.once('uncaughtException', function (e) {
     debugger;
-    _suman.logError(chalk.magenta('Suman runner "uncaughtException" event:') + " \n " + chalk.bold(suman_utils_1.default.getCleanErrorString(e)));
+    _suman.logError(chalk.magenta('Suman runner "uncaughtException" event:') + " \n " + chalk.bold(su.getCleanErrorString(e)));
     process.exit(1);
 });
 process.on('message', function (data) {
@@ -178,16 +177,16 @@ exports.findTestsAndRunThem = function (runObj, runOnce, $order) {
         resultBroadcaster.emit(String(suman_events_1.events.ERRORS_ONLY_OPTION));
     }
     delete process.env.SUMAN_EXTRANEOUS_EXECUTABLE;
-    var projectRoot = _suman.projectRoot || suman_utils_1.default.findProjectRoot(process.cwd());
+    var projectRoot = _suman.projectRoot || su.findProjectRoot(process.cwd());
     runnerObj.handleBlocking = make_handle_blocking_1.default(mapValues($order, function (val) {
         val.testPath = path.resolve(projectRoot + '/' + val.testPath);
         return val;
     }));
     process.nextTick(function () {
         var args = fnArgs(runOnce);
-        var ret = runOnce.apply(null, integrant_injector_1.default(args));
+        var ret = runOnce.apply(null, integrant_injector_1.default(args, null));
         if (ret.dependencies) {
-            if (suman_utils_1.default.isObject(ret.dependencies)) {
+            if (su.isObject(ret.dependencies)) {
                 runnerObj.depContainerObj = ret.dependencies;
             }
             else {
